Move socket to module, emit current users in room
